container: Move container_under_pointer here from handlers.
