+++ include/container.h
@@ -5,6 +5,8 @@ typedef struct sway_container swayc_t;
 
 #include "layout.h"
 
+extern struct wlc_origin mouse_origin;
+
 enum swayc_types{
 	C_ROOT,
 	C_OUTPUT,
@@ -98,6 +100,8 @@ swayc_t *swayc_by_name(const char *name);
 swayc_t *swayc_active_output(void);
 swayc_t *swayc_active_workspace(void);
 swayc_t *swayc_active_workspace_for(swayc_t *view);
+// set focus to current pointer location and return focused container
+swayc_t *container_under_pointer(void);
 
 // Container information
